backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
djangorestframework-simplejwt==5.3.1
